move lowerRegexpASCII and bytesToLowerASCII into sharable package
