--- cmd/searcher/search/lower_amd64.go
@@ -1,4 +1,0 @@
-package search
-
-// implemented in assembly, see lower_amd64.s
-func bytesToLowerASCII(dst, src []byte)
