--- cmd/searcher/search/lower_other.go
@@ -1,6 +1,0 @@
-//go:build !amd64
-// +build !amd64
-
-package search
-
-var bytesToLowerASCII = bytesToLowerASCIIgeneric
